requirements.txt
fastapi
uvicorn
dnspython
httpx[http2,socks]
jinja2
python-multipart
